+++ src/Stripe.net/Entities/Plans/Plan.cs
@@ -20,7 +20,7 @@ namespace Stripe
         public string AggregateUsage { get; set; }
 
         [JsonProperty("amount")]
-        public int? Amount { get; set; }
+        public long? Amount { get; set; }
 
         [JsonProperty("billing_scheme")]
         public string BillingScheme { get; set; }
@@ -41,7 +41,7 @@ namespace Stripe
         public string Interval { get; set; }
 
         [JsonProperty("interval_count")]
-        public int IntervalCount { get; set; }
+        public long IntervalCount { get; set; }
 
         [JsonProperty("livemode")]
         public bool Livemode { get; set; }
@@ -83,7 +83,7 @@ namespace Stripe
         public PlanTransformUsage TransformUsage { get; set; }
 
         [JsonProperty("trial_period_days")]
-        public int? TrialPeriodDays { get; set; }
+        public long? TrialPeriodDays { get; set; }
 
         [JsonProperty("usage_type")]
         public string UsageType { get; set; }
