+++ src/Stripe.net/Services/ListOptions.cs
@@ -8,7 +8,7 @@ namespace Stripe
         /// A limit on the number of objects to be returned. Limit can range between 1 and 100 items.
         /// </summary>
         [JsonProperty("limit")]
-        public int? Limit { get; set; }
+        public long? Limit { get; set; }
 
         /// <summary>
         /// A cursor for use in pagination. starting_after is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include starting_after=obj_foo in order to fetch the next page of the list.
